remove sidebar controller
