+++ KioskApplication/src/main/java/controller/MapBuilderController.java
@@ -1,10 +1,21 @@
 package controller;
 
 import database.objects.Edge;
+import database.util.CSVFileUtil;
+import entity.MapEntity;
+import javafx.fxml.FXML;
 import javafx.geometry.Point2D;
+import javafx.scene.control.Tab;
+import javafx.scene.control.TabPane;
 import utility.Node.NodeFloor;
 
 public class MapBuilderController extends ScreenController {
+
+    @FXML
+    private TabPane builderTabPane;
+    private Tab nodeTab;
+    private Tab edgeTab;
+    private Tab databaseTab;
 
     MapBuilderController(MainWindowController parent, MapController map) {
         super(parent, map);
@@ -34,5 +45,55 @@ public class MapBuilderController extends ScreenController {
     @Override
     public void resetScreen() {
         getMapController().setAnchor(0, 200, 0, 0);
+    }
+
+    @FXML
+    void onReadClicked() {
+        // TODO implement this better
+        // Load nodes
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapAnodes.csv"));
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapBnodes.csv"));
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapCnodes.csv"));
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapDnodes.csv"));
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapEnodes.csv"));
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapFnodes.csv"));
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapGnodes.csv"));
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapHnodes.csv"));
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapInodes.csv"));
+        CSVFileUtil.readNodesCSV(getClass().getResourceAsStream("/csv/MapWnodes.csv"));
+
+        // Load edges
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapAedges.csv"));
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapBedges.csv"));
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapCedges.csv"));
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapDedges.csv"));
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapEedges.csv"));
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapFedges.csv"));
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapGedges.csv"));
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapHedges.csv"));
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapIedges.csv"));
+        CSVFileUtil.readEdgesCSV(getClass().getResourceAsStream("/csv/MapWedges.csv"));
+
+        MapEntity.getInstance().readAllFromDatabase();
+
+        getMapController().reloadDisplay();
+    }
+
+    @FXML
+    void onSaveClicked() {
+        // TODO Implement SaveCSV with different team letters
+        /*
+        try {
+            URI mapINodes = new URI(getClass().getResource("/csv/MapInodes.csv").toString());
+            CSVFileUtil.writeNodesCSV(mapINodes.getPath(), false);
+
+            URI mapWNodes = new URI(getClass().getResource("/csv/MapWnodes.csv").toString());
+            CSVFileUtil.writeNodesCSV(mapWNodes.getPath(), true);
+
+            URI mapIEdges = new URI(getClass().getResource("/csv/MapIedges.csv").toString());
+            CSVFileUtil.writeEdgesCSV(mapIEdges.getPath(), false);
+        } catch (URISyntaxException e) {
+            e.printStackTrace();
+        }*/
     }
 }
